+++ src/components/layout/Header.jsx
@@ -7,10 +7,10 @@ import PeopleIcon from '@mui/icons-material/People';
 import FavoriteIcon from '@mui/icons-material/Favorite';
 import AddShoppingCartIcon from '@mui/icons-material/AddShoppingCart';
 import LogoutIcon from '@mui/icons-material/Logout';
-import Message from "../Message";
 import Banner from "../Banner";
+import AdMessage from "../AdMessage";
 import NotificationsIcon from '@mui/icons-material/Notifications';
-import SwipeableEdgeDrawer from "../AlarmDrawer"
+import SwipeableEdgeDrawer from "../MyMessage"
 import ExitToAppIcon from '@mui/icons-material/ExitToApp';
 import VpnKeyIcon from '@mui/icons-material/VpnKey';
 //styleHeader.HeaderClass:헤더 전체
@@ -66,7 +66,7 @@ const Header = () => {
 
     return (
         <>
-        <Message ><Banner /></Message>
+        <Banner ><AdMessage /></Banner>
         <div className={styleHeader.HeaderClass}>
             <h1 id={styleHeader['logo']}
                 onClick={() => goToMenu('/')}
